Fix return to task when single task activity is brought to front
When a single task activity launched on top of home is relaunched
from another flow the return to task for the above activity is not
set correctly.

Adding a check and setting the returnto task to same as the
relaunched single task activity reurnto value ensures correct
activity is resumed.

CRs-Fixed: 830806
Change-Id: I3acf3e301ac3acd4e7aa69bdaf1cf7281b38c907
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>